expose utility functions
* args2cmdline and open_fifo_write are needed by awx/tower. Move them to
be utility functions instead of private runner_config functions.
